+++ tests/envs/test_pybullet_handover.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from multitask_personalization.envs.pybullet.pybullet_handover import (
+from multitask_personalization.envs.pybullet.pybullet_tasks import (
     PyBulletHandoverTask,
 )
