+++ tests/methods/calibration/test_pybullet_handover_calibrator.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from multitask_personalization.envs.pybullet.pybullet_handover import (
+from multitask_personalization.envs.pybullet.pybullet_tasks import (
     PyBulletHandoverTask,
 )
 from multitask_personalization.methods.calibration.pybullet_handover_calibrator import (
